Merge dev to master

+++ src/main/java/aptech/project/educhain/domain/useCases/blogs/BlogCommentUseCases/DeleteCommentUseCase.java
@@ -1,5 +1,6 @@
 package aptech.project.educhain.domain.useCases.blogs.BlogCommentUseCases;
 
+import aptech.project.educhain.data.entities.blogs.BlogComment;
 import aptech.project.educhain.data.repositories.blogs.BlogCommentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -11,8 +12,14 @@ public class DeleteCommentUseCase {
 
     public boolean execute(Integer id) {
         try {
-            blogCommentRepository.deleteById(id);
-            return true;
+            BlogComment blogComment = blogCommentRepository.findById(id).get();
+            if (blogComment != null){
+                blogCommentRepository.deleteById(id);
+                return true;
+            }else{
+                return false;
+            }
+
         }catch(Exception ex){
             ex.getMessage();
         }
